[refactor] Move typebuilders into dsl namespace
